backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
pydantic
pydantic-settings
python-dotenv
websockets
